refactor blog post list and switch theme
