--- src/components/NewPost/styles.ts
@@ -1,1 +1,0 @@
-export {}
