--- src/components/SwitchTheme/index.ts
@@ -1,1 +1,0 @@
-export * from './SwitchTheme'
