--- src/components/Weather/index.ts
@@ -1,1 +1,0 @@
-export { WeatherConditions } from './weather'
