--- src/components/Weather/styled.ts
@@ -1,16 +1,0 @@
-import { styled } from 'src/ui/theme'
-
-export const Weather = styled('span', {
-  all: 'unset',
-  display: 'flex',
-  alignItems: 'flex-start',
-  justifyContent: 'flex-start',
-  flexFlow: 'row wrap',
-  lineHeight: '1.6',
-  '--fluid-type-max': '1rem',
-
-  div: {
-    display: 'flex',
-    alignItems: 'center',
-  },
-})
