--- src/domains/blog/article-card/index.ts
@@ -1,1 +1,0 @@
-export { ArticleCard } from './article-card.component'
