--- src/domains/blog/blog-posts/index.ts
@@ -1,1 +1,0 @@
-export { BlogPosts } from './blog-posts.component'
