--- src/domains/blog/blog-posts/styles.ts
@@ -1,9 +1,0 @@
-import { styled } from 'src/ui/theme'
-
-export const ArticlesContainer = styled('ul', {
-  display: 'flex',
-  flexDirection: 'column',
-  padding: '3rem 0',
-  listStyle: 'none',
-  margin: 0,
-})
