+++ src/domains/blog/list/styles.module.css
@@ -1,0 +1,7 @@
+.articlesContainer {
+  display: flex;
+  flex-direction: column;
+  padding: 3rem 0;
+  list-style: none;
+  margin: 0;
+}
