+++ src/domains/blog/posts/index.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { TfiPencil } from 'react-icons/tfi'
 import { getLocaleDate } from 'utils/get-locale-date'
-import { CardContainer, DateTimeRead, PostTitle } from './styles'
+import styles from './styles.module.css'
 
 const ArticleCard = ({ article }: any) => {
   const { slug, title, publishedAt } = article.attributes
@@ -9,15 +9,15 @@ const ArticleCard = ({ article }: any) => {
   const { localeDate } = getLocaleDate(publishedAt, 'pt-BR')
 
   return (
-    <CardContainer>
-      <DateTimeRead>
+    <li className={styles.cardContainer}>
+      <p className={styles.dateTimeRead}>
         <TfiPencil size={17} />
         {localeDate}
-      </DateTimeRead>
+      </p>
       <Link href={`/blog/article/${slug}`} passHref suppressHydrationWarning>
-        <PostTitle>{title}</PostTitle>
+        <p className={styles.postTitle}>{title}</p>
       </Link>
-    </CardContainer>
+    </li>
   )
 }
